+++ changelog.md
@@ -6,6 +6,9 @@ All important changes to this plugin will be documented in this file.
 ### Changed
 - Adjusted plugin to work with ilias 6. MUMIE Task no longer supports Ilias versions prior to 6.0
 
+### Fixed
+- Fixed an error in MUMIE server form, where whitespace around the URL prefix could cause an error 
+
 ## [v1.2.0] - 2021-04-29
 ### Fixed
 - Editing a MUMIE Task no longer changes the selected problem.